+++ custom_components/omniremote/const.py
@@ -7,7 +7,7 @@ from typing import Any
 import uuid
 
 DOMAIN = "omniremote"
-VERSION = "1.10.12"
+VERSION = "1.10.13"
 
 # Debug flag - set to True for verbose logging
 DEBUG = True